fix to enable code (parallel workers/env in ppo agent) to run on MacOS using Python>=3.8.
note: lambda functions passed to threads or forked processes
do not work when the multiprocessing start method is set to 'spawn' by default.
In MacOS, the default prior to Python 3.8 was 'fork' (which works fine). however,
this was changed to 'spawn' in Python 3.8.

second note: linux does not need this fix because the default multiprocessing start
method is 'fork' and has not been changed.

 On branch main
 Your branch is up to date with 'origin/main'.

 Changes to be committed:
	modified:   deep_rl/component/task.py