*reward normalizer* added reward normalizer implementation and added reward normalization in all (meta-)continualworld experiments across single agent and shell experiments runner.
*update: training iteration reward logging.*
*deep_rl/agent/PPO_agent.py, deep_rl/utils/trainer_ll.py, and deep_rl/utils/trainer_shell.py*
refactoring and logging of average episode return across all episodes within a single iteration.
we can now log both the last episode return in an iteration (which was the existing code), as well
as the average returns across all episodes within a single iteration.

*agent evaluation update: deep_rl/agent/BaseAgent.py*
added code to return evaluation performance as either success rate for (meta-)continualworld
environments or rewards for other environments.

 Changes to be committed:
	modified:   deep_rl/agent/BaseAgent.py
	modified:   deep_rl/agent/PPO_agent.py
	modified:   deep_rl/utils/normalizer.py
	modified:   deep_rl/utils/trainer_ll.py
	modified:   deep_rl/utils/trainer_shell.py
	modified:   run_ll.py
	modified:   run_shell.py
	modified:   run_shell_dist.py