implemented code to support initialising a new mask from scratch for a previously unseen task and then training the mask, OR use a weighted linear combination of previous task mask and new task mask to bootstrap training on the current task. note, during the task training, the weight co-efficients are also trained (updated by backprop). although the masks previous tasks do not get updated during training on the current task, the co-efficients used to linearly combine them is updated.
In essence, given task A, B and C, with previous training on A and B with their
corresponding mask, when training on task C, the parameters that will be updated
include:
the randomly initlaised mask for task C, co-efficients for task A mask, co-efficient
for task B mask, and co-efficient for task C mask. The masks for task A and B respectively
will not be updated.

 On branch shell
 Your branch is up-to-date with 'origin/shell'.

 Changes to be committed:
	modified:   deep_rl/agent/PPO_agent.py
	modified:   deep_rl/shell_modules/mmn/ssmask_utils.py