added a second version of continuous mask, where the gradient of the loss w.r.t to the raw (mask) score is actually computed (instead of passed through in discrete mask and original continuous mask version).
The gradient was originally skipped (or passed through or straight through estimator)
due to discrete/binary setting where the gradient will be 0, hence we pass through
the gradient from downstream operation to upstream operations of a supermask linear layer.

note: when the second version was tested (in single agent setting using meta/cont-
inualworld environment), it did not work as well as the original/first version where
the gradient for the raw (mask) score are copies of the gradient with respect to the
processed continuous mask (downstream operation).
Hence, the default choice of continuous mask is still the one earlier implemented

 On branch dev
 Your branch is up to date with 'origin/dev'.

 Changes to be committed:
	modified:   deep_rl/shell_modules/mmn/ssmask_utils.py